move all to index
